move `and` and `or` conditions to `conditions` package
Having this inside of `ArchCondition` is a little inconsistent, since all the other conditions, including `never` or `containsOnly`, reside inside the subpackage `conditions`. Also, these conditions can then not share package private classes with the other conditions.
